Revert "Models: split enums and types"
This reverts commit debb0594cd874385c717983a64f17cb182654e90.
