+++ src/index.ts
@@ -1,7 +1,6 @@
 import UbirchVerification from './verification';
 import UbirchVerificationWidget from './widget';
 import UbirchFormUtils from './form-utils';
-import * as types from './models/types';
-import * as enums from './models/enums';
+import * as models from './models/models';
 
-export { UbirchVerification, UbirchVerificationWidget, UbirchFormUtils, types, enums };
+export { UbirchVerification, UbirchVerificationWidget, UbirchFormUtils, models };
